requirements.txt
# Core ML and NLP libraries
torch>=2.1.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
deepspeed>=0.12.0

# Hugging Face Hub integration
huggingface_hub>=0.17.0

# Text processing and evaluation
textstat>=0.7.3
detoxify>=0.5.2
openai>=1.3.0

# Data manipulation and analysis
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Data downloading
kaggle>=1.5.16

# Development tools
jupyter>=1.0.0